+++ app/src/main/java/eu/kanade/tachiyomi/data/track/komga/KomgaApi.kt
@@ -63,7 +63,7 @@ class KomgaApi(private val client: OkHttpClient) {
                 track.apply {
                     cover_url = "$url/thumbnail"
                     tracking_url = url
-                    total_chapters = progress.maxNumberSort.toInt()
+                    total_chapters = progress.maxNumberSort.toLong()
                     status = when (progress.booksCount) {
                         progress.booksUnreadCount -> Komga.UNREAD
                         progress.booksReadCount -> Komga.COMPLETED
