+++ app/src/main/java/eu/kanade/tachiyomi/data/track/suwayomi/TachideskApi.kt
@@ -11,6 +11,8 @@ import eu.kanade.tachiyomi.network.PUT
 import eu.kanade.tachiyomi.network.awaitSuccess
 import eu.kanade.tachiyomi.network.parseAs
 import eu.kanade.tachiyomi.util.system.withIOContext
+import java.nio.charset.Charset
+import java.security.MessageDigest
 import kotlinx.serialization.json.Json
 import okhttp3.Credentials
 import okhttp3.Dns
@@ -20,8 +22,6 @@ import okhttp3.OkHttpClient
 import uy.kohesive.injekt.Injekt
 import uy.kohesive.injekt.api.get
 import uy.kohesive.injekt.injectLazy
-import java.nio.charset.Charset
-import java.security.MessageDigest
 
 class TachideskApi {
     private val json: Json by injectLazy()
@@ -61,7 +61,7 @@ class TachideskApi {
             cover_url = "$url/thumbnail"
             summary = manga.description.orEmpty()
             tracking_url = url
-            total_chapters = manga.chapterCount.toInt()
+            total_chapters = manga.chapterCount
             publishing_status = manga.status
             last_chapter_read = manga.lastChapterRead?.chapterNumber ?: 0F
             status = when (manga.unreadCount) {
