+++ app/src/main/java/eu/kanade/tachiyomi/widget/preference/TrackLogoutDialog.kt
@@ -2,10 +2,6 @@ package eu.kanade.tachiyomi.widget.preference
 
 import android.app.Dialog
 import android.os.Bundle
-import eu.kanade.tachiyomi.R
-import yokai.i18n.MR
-import yokai.util.lang.getString
-import dev.icerock.moko.resources.compose.stringResource
 import eu.kanade.tachiyomi.data.track.TrackManager
 import eu.kanade.tachiyomi.data.track.TrackService
 import eu.kanade.tachiyomi.ui.base.controller.DialogController
@@ -15,12 +11,14 @@ import eu.kanade.tachiyomi.util.view.setNegativeButton
 import eu.kanade.tachiyomi.util.view.setPositiveButton
 import uy.kohesive.injekt.Injekt
 import uy.kohesive.injekt.api.get
+import yokai.i18n.MR
+import yokai.util.lang.getString
 
 class TrackLogoutDialog(bundle: Bundle? = null) : DialogController(bundle) {
 
-    private val service = Injekt.get<TrackManager>().getService(args.getInt("key"))!!
+    private val service = Injekt.get<TrackManager>().getService(args.getLong("key"))!!
 
-    constructor(service: TrackService) : this(Bundle().apply { putInt("key", service.id) })
+    constructor(service: TrackService) : this(Bundle().apply { putLong("key", service.id) })
 
     override fun onCreateDialog(savedViewState: Bundle?): Dialog {
         val serviceName = activity!!.getString(service.nameRes())
